Cast the enum to satisfy CI

+++ src/opts.rs
@@ -8,7 +8,7 @@ use crate::raw;
 pub fn strict_hash_verification(enabled: bool) {
     let error = unsafe {
         raw::git_libgit2_opts(
-            raw::GIT_OPT_ENABLE_STRICT_HASH_VERIFICATION,
+            raw::GIT_OPT_ENABLE_STRICT_HASH_VERIFICATION as libc::c_int,
             if enabled { 1 } else { 0 } as libc::c_int,
         )
     };
